вынес работу с датами в одельный утилитарный модуль
